% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexPartsR.R
\name{convexParts}
\alias{convexParts}
\title{Title}
\usage{
convexParts(vertices, faces, mesh = NULL, triangulate = FALSE)
}
\arguments{
\item{vertices}{x}

\item{faces}{x}

\item{mesh}{x}

\item{triangulate}{x}
}
\value{
x
}
\description{
Title
}
\examples{
#
}
