% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someMeshes.R
\name{cyclideMesh}
\alias{cyclideMesh}
\title{Cyclide mesh}
\usage{
cyclideMesh(a, c, mu, nu = 90, nv = 40, rgl = TRUE)
}
\arguments{
\item{a, c, mu}{cyclide parameters, positive numbers such that
\code{c < a} and \code{c < mu}}

\item{nu, nv}{numbers of subdivisions, integers (at least 3)}

\item{rgl}{Boolean, whether to return a \strong{rgl} mesh}
}
\value{
A triangle \strong{rgl} mesh (class \code{mesh3d}) if
  \code{rgl=TRUE}, otherwise a \code{cgalMesh} list (vertices, faces,
  and normals).
}
\description{
Triangle mesh of a Dupin cyclide.
}
\examples{
library(MeshesOperations)
library(rgl)
mesh <- cyclideMesh(a = 0.97, c = 0.32, mu = 0.56)
open3d(windowRect = c(50, 50, 562, 562))
view3d(0, 0, zoom = 0.75)
shade3d(mesh, color = "chartreuse")
wire3d(mesh)
}
