% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remeshing.R
\name{isotropicRemesh}
\alias{isotropicRemesh}
\title{Isotropic remeshing}
\usage{
isotropicRemesh(
  vertices,
  faces,
  targetEdgeLength,
  iterations = 1,
  relaxSteps = 1,
  normals = FALSE
)
}
\arguments{
\item{vertices}{a numeric matrix with three columns}

\item{faces}{either an integer matrix (each row provides the vertex indices
of the corresponding face) or a list of integer vectors, each one
providing the vertex indices of the corresponding face}

\item{targetEdgeLength}{positive number, the target edge length of the 
remeshed mesh}

\item{iterations}{number of iterations, a positive integer}

\item{relaxSteps}{number of relaxation steps, a positive integer}

\item{normals}{Boolean, whether to compute the vertex normals of the 
output mesh}
}
\value{
A triangle mesh represented as the output of the 
  \code{\link{Mesh}} function.
}
\description{
Isotropically remesh a mesh.
}
\examples{
library(MeshesOperations)
library(rgl)
}
