% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{meshArea}
\alias{meshArea}
\title{Mesh area}
\usage{
meshArea(mesh)
}
\arguments{
\item{mesh}{xxx}
}
\value{
A number, the surface area of the mesh.
}
\description{
Computes the surface area a mesh.
}
\examples{
library(MeshesOperations)
R <- 4; r <- 2
mesh <- torusMesh(R, r)
meshArea(mesh)
# true area of the torus: 
4 * pi^2 * R * r
}
