% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinkowskiSum.R
\name{MinkowskiSum}
\alias{MinkowskiSum}
\title{Minkowski sum of two meshes}
\usage{
MinkowskiSum(mesh1, mesh2, normals = FALSE)
}
\arguments{
\item{mesh1, mesh2}{xxx}

\item{normals}{Boolean, whether to compute the vertex normals of the 
output mesh}
}
\value{
xxx
}
\description{
Returns the mesh defined as the Minkowski sum of the 
  two input meshes.
}
\examples{
library(MeshesOperations)
library(rgl)
library(Rvcg)
# Stanford bunny (light version)
mesh1 <- readMeshFile(
  system.file("extdata", "bunny.off", package = "MeshesOperations")
)
mesh2 <- vcgSphere(subdivision = 2, normals = FALSE)
mesh <- MinkowskiSum(mesh1, mesh2)
rglmesh <- toRGL(mesh)
open3d(windowRect = c(50, 50, 900, 500))
view3d(0, 0, zoom = 0.8)
shade3d(rglmesh, color = "maroon")
}
