% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-octahedraCompound.R
\docType{data}
\name{octahedraCompound}
\alias{octahedraCompound}
\title{Compound of five octahedra}
\format{
A list with three fields: the field \code{meshes} is a list of five
  elements, each one representing an octahedron by a list with two elements,
  the vertices and the faces; the field \code{rglmeshes} is the list of
  the five corresponding \strong{rgl} meshes; the field \code{gmpmeshes} is
  the same as \code{meshes} except that the vertices are \strong{gmp}
  rational numbers.
}
\usage{
octahedraCompound
}
\description{
Five octahedra in a pretty configuration. Each octahedron
  is centered at the origin.
}
\keyword{datasets}
