% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pentagrammicPrism.R
\docType{data}
\name{pentagrammicPrism}
\alias{pentagrammicPrism}
\title{A mesh of a pentagrammic prism}
\format{
A \code{cgalMesh} list (\code{vertices}, \code{edges}, \code{faces}).
}
\usage{
pentagrammicPrism
}
\description{
A \code{cgalMesh} list representing a pentagrammic prism; it
  has 20 vertices, 10 triangular faces, 10 rectangular faces and two
  pentagonal faces.
}
\keyword{datasets}
