% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tetrahedraCompound.R
\docType{data}
\name{tetrahedraCompound}
\alias{tetrahedraCompound}
\title{Compound of five tetrahedra}
\format{
A list with three fields: the field \code{meshes} is a list of five
  elements, each one representing a tetrahedron by a list with two elements,
  the vertices and the faces; the field \code{rglmeshes} is the list of
  the five corresponding \strong{rgl} meshes; the field \code{gmpmeshes} is
  the same as \code{meshes} except that the vertices are \strong{gmp}
  rational numbers.
}
\usage{
tetrahedraCompound
}
\description{
Five tetrahedra in a pretty configuration. Each tetrahedron
  is centered at the origin.
}
\keyword{datasets}
