% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{clipMesh}
\alias{clipMesh}
\title{Clip a mesh}
\usage{
clipMesh(mesh, clipper, clipVolume, normals = FALSE)
}
\arguments{
\item{mesh}{a mesh given either as a list containing (at least) the fields 
\code{vertices} and \code{faces}, otherwise a \strong{rgl} mesh 
(i.e. a \code{mesh3d} object)}

\item{clipper}{a mesh given either as a list containing (at least) the fields 
\code{vertices} and \code{faces}, otherwise a \strong{rgl} mesh 
(i.e. a \code{mesh3d} object); if \code{clipVolume=TRUE}, this mesh will be 
modified: it will be refined with the intersection with \code{mesh}}

\item{clipVolume}{Boolean, whether the clipping has to be done on the volume 
bounded by \code{mesh} rather than on its surface (i.e. \code{mesh} will be 
kept closed if it is closed)}

\item{normals}{Boolean, whether to compute the vertex normals of the 
output mesh(es)}
}
\value{
If \code{clipVolume=FALSE}, a triangle mesh represented as the output of the 
  \code{\link{Mesh}} function, otherwise a list of two such triangle meshes: the 
  clipped mesh in the field \code{"mesh"} and the modified clipping mesh in the 
  field \code{"clipper"}.
}
\description{
Clip a mesh to the volume bounded by another mesh.
}
\note{
If \code{clipVolume=TRUE}, the mesh to be clipped (\code{"mesh"}) must be 
  without self-intersection.
}
\examples{
library(MeshesOperations)
library(rgl)
library(Rvcg)
mesh    <- cube3d()
clipper <- scale3d(vcgSphere), sqrt(2), sqrt(2), sqrt(2))
clippedMesh <- clipMesh(mesh, clipper, clipVolume = TRUE)
open3d(windowRect = c(50, 50, 950, 500))
mfrow3d(1, 2)
view3d(zoom = 0.9)
shade3d(toRGL(clippedMesh[["mesh"]]), color = "purple")
next3d()
view3d(zoom = 0.9)
shade3d(toRGL(clippedMesh[["clipper"]]), color = "khaki")
}
