% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinkowskiSum.R
\name{MinkowskiSum}
\alias{MinkowskiSum}
\title{Minkowski sum of two meshes}
\usage{
MinkowskiSum(mesh1, mesh2, triangulate = TRUE, normals = FALSE)
}
\arguments{
\item{mesh1, mesh2}{xxx}

\item{triangulate}{Boolean, whether to triangulate the output mesh (note 
that it is not necessarily triangle when the two input meshes are triangle)}

\item{normals}{Boolean, whether to compute the vertex normals of the 
output mesh}
}
\value{
xxx
}
\description{
Returns the mesh defined as the Minkowski sum of the 
  two input meshes.
}
\examples{
library(MeshesOperations)
library(rgl)
library(Rvcg)
mesh1 <- octahedron3d()
mesh2 <- vcgSphere(subdivision = 2, normals = FALSE)
mesh <- MinkowskiSum(mesh1, mesh2, normals = TRUE)
rglmesh <- toRGL(mesh)
open3d(windowRect = c(50, 50, 562, 562))
view3d(30, 30, zoom = 0.8)
shade3d(rglmesh, color = "maroon")
}
